pyproject.toml
[tool.poetry]
name = "final-challenge"
version = "0.1.0"
description = ""
authors = ["Lucas Torres <lucas.dionatan@mercai.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "*"}
sqlmodel = "*"
pydantic-settings = "*"
psycopg2-binary = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
